ASE: resampler2: import factor-2 resampler from Stefan Westerfeld
Import resampler from:
  https://github.com/tim-janik/beast/blob/master/bse/bseresampler.hh
  https://github.com/tim-janik/beast/blob/master/bse/bseresampler.cc

Code originally by:
  Stefan Westerfeld <stefan@space.twc.de>

Squashed commits since 2006-09-16:

Added Resampler2::find_precision_for_bits to the API, and a
precision enum -> string conversion function.

The guarantee that the upsampler delay is even was removed from the API.
The reason is that we can have a shorter stream delay when the upsampler
delay is odd.

Removed unnecessary delay. The stream delay for both, upsampling and
downsampling should be truely minimal now.

Added coefficient set for linear interpolation. It will probably sound
bad, but its useful for testing.

Improved comments on delay compensation.

Added a virtual function Resampler2::delay(),
which can be used for compensation of the resampler introduced delay.

Added factor 2 resampling code (up- and downsampling), which can use SSE
instructions when available. The implementation was tuned for speed.

Signed-off-by: Stefan Westerfeld <stefan@space.twc.de>
Signed-off-by: Tim Janik <timj@gnu.org>